get_format_options loads from saved version json
